+++ tests/publisher_test.py
@@ -23,8 +23,10 @@ import unittest
 
 import mock
 
+from perfkitbenchmarker import disk
 from perfkitbenchmarker import publisher
 from perfkitbenchmarker import sample
+from perfkitbenchmarker import vm_util
 
 
 class PrettyPrintStreamPublisherTestCase(unittest.TestCase):
@@ -106,6 +108,7 @@ class BigQueryPublisherTestCase(unittest.TestCase):
   def setUp(self):
     p = mock.patch(publisher.__name__ + '.vm_util', spec=publisher.vm_util)
     self.mock_vm_util = p.start()
+    publisher.vm_util.NamedTemporaryFile = vm_util.NamedTemporaryFile
     self.mock_vm_util.GetTempDir.return_value = tempfile.gettempdir()
     self.addCleanup(p.stop)
 
@@ -154,6 +157,7 @@ class CloudStoragePublisherTestCase(unittest.TestCase):
   def setUp(self):
     p = mock.patch(publisher.__name__ + '.vm_util', spec=publisher.vm_util)
     self.mock_vm_util = p.start()
+    publisher.vm_util.NamedTemporaryFile = vm_util.NamedTemporaryFile
     self.mock_vm_util.GetTempDir.return_value = tempfile.gettempdir()
     self.addCleanup(p.stop)
 
@@ -219,3 +223,68 @@ class SampleCollectorTestCase(unittest.TestCase):
     samples = [tuple(self.sample)]
     self.instance.AddSamples(samples, self.benchmark, self.benchmark_spec)
     self._VerifyResult()
+
+
+class DefaultMetadataProviderTestCase(unittest.TestCase):
+
+  def setUp(self):
+    p = mock.patch(publisher.__name__ + '.FLAGS')
+    self.mock_flags = p.start()
+    self.mock_flags.configure_mock(metadata=[],
+                                   num_striped_disks=1)
+    self.addCleanup(p.stop)
+
+    p = mock.patch(publisher.__name__ + '.version',
+                   VERSION='v1')
+    p.start()
+    self.addCleanup(p.stop)
+
+    self.mock_spec = mock.MagicMock(cloud='GCP',
+                                    zones=['us-central1-a'],
+                                    machine_type='n1-standard-1',
+                                    image='ubuntu-14-04')
+
+    self.default_meta = {'perfkitbenchmarker_version': 'v1',
+                         'cloud': self.mock_spec.cloud,
+                         'zones': 'us-central1-a',
+                         'machine_type': self.mock_spec.machine_type,
+                         'image': self.mock_spec.image,
+                         'num_striped_disks': 1}
+
+  def _RunTest(self, spec, expected, input_metadata=None):
+    input_metadata = input_metadata or {}
+    instance = publisher.DefaultMetadataProvider()
+    result = instance.AddMetadata(input_metadata, self.mock_spec)
+    self.assertIsNot(input_metadata, result,
+                     msg='Input metadata was not copied.')
+    self.assertDictEqual(expected, result)
+
+  def testAddMetadata_ScratchDiskUndefined(self):
+    del self.mock_spec.scratch_disk
+    meta = self.default_meta.copy()
+    meta.pop('num_striped_disks')
+    self._RunTest(self.mock_spec, meta)
+
+  def testAddMetadata_NoScratchDisk(self):
+    self.mock_spec.scratch_disk = False
+    meta = self.default_meta.copy()
+    meta.pop('num_striped_disks')
+    self._RunTest(self.mock_spec, meta)
+
+  def testAddMetadata_WithScratchDisk(self):
+    self.mock_spec.configure_mock(scratch_disk=True,
+                                  scratch_disk_size=20,
+                                  scratch_disk_type='remote_ssd')
+    expected = self.default_meta.copy()
+    expected.update(scratch_disk_size=20, scratch_disk_type='remote_ssd')
+    self._RunTest(self.mock_spec, expected)
+
+  def testAddMetadata_PIOPS(self):
+    self.mock_spec.configure_mock(scratch_disk=True,
+                                  scratch_disk_size=20,
+                                  scratch_disk_type=disk.PIOPS,
+                                  scratch_disk_iops=1000)
+    expected = self.default_meta.copy()
+    expected.update(scratch_disk_size=20, scratch_disk_type=disk.PIOPS,
+                    scratch_disk_iops=1000)
+    self._RunTest(self.mock_spec, expected)