+++ tests/vm_util_test.py
@@ -14,6 +14,7 @@
 
 """Tests for perfkitbenchmarker.vm_util."""
 
+import subprocess
 import unittest
 
 import mock
@@ -65,6 +66,27 @@ class ShouldRunOnInternalIpAddressTestCase(unittest.TestCase):
   def testReachable_Unreachable(self):
     self._RunTest(
         False, vm_util.IpAddressSubset.REACHABLE, False)
+
+
+class IssueCommandTestCase(unittest.TestCase):
+
+  def testTimeoutNotReached(self):
+    _, _, retcode = vm_util.IssueCommand(['sleep', '2s'])
+    self.assertEqual(retcode, 0)
+
+  def testTimeoutReached(self):
+    _, _, retcode = vm_util.IssueCommand(['sleep', '2s'], timeout=1)
+    self.assertEqual(retcode, -9)
+
+  def testNoTimeout(self):
+    _, _, retcode = vm_util.IssueCommand(['sleep', '2s'], timeout=None)
+    self.assertEqual(retcode, 0)
+
+  def testNoTimeout_ExceptionRaised(self):
+    with mock.patch('subprocess.Popen', spec=subprocess.Popen) as mock_popen:
+      mock_popen.return_value.communicate.side_effect = KeyboardInterrupt()
+      with self.assertRaises(KeyboardInterrupt):
+        vm_util.IssueCommand(['sleep', '2s'], timeout=None)
 
 
 if __name__ == '__main__':